Update the artifact production process
Fully divorce artifact creation from the Lambda image. We switch to
mounting the build_artifact.sh into the container as part of the
build_deployment_package service and move installing zip into that
script. We also remove the BUILD_FILE_NAME environment variable from
the Docker image and house it as part of the Docker Compose
configuration.

These changes better align the base image as a minimal AWS Lambda image
suited for use on its own. Building a deployment artifact now leverages
this image instead of being part of the image.