Remove the release.yml workflow
Work toward a fancy automated release process can be pushed off until
after the initial implementation is in place. This will allow
appropriate testing of an otherwise functional implementation so it is
appropriately isolated. In the interim releases can be managed by
manually creating a release and uploading the artifacts generated from
the merge commit's Actions run.