Use a multi-stage build for the Lambda image
Part of this project's goals is to support the option of running a
bespoke image as a Lambda and not just using deployment packages. Since
we care about the created image due to this we should build it similar
to how we would build other Docker images. Thus we switch to a multi-
stage build so that the only things included in the final image are
what is absolutely necessary to run the Lambda (per the dependency
files). With this configuration we saw space savings of almost 10% by
switching to a multi-stage build.