Ensure pipenv uses the system Python instead of a venv
The default behavior of pipenv is to create a virtual environment to
use for installing dependencies. Our configuration does not leverage
this venv since we are simply using pipenv to manage dependencies and
need all dependencies installed into the Lambda's task root directory.
This default behavior can create issues with our configuration because
if one of the core packages seeded into that venv (pip, setuptools, or
wheel) is the same version as a dependency in the lockfile then it will
not be installed into the Lambda's task root directory for later use..