Add functionality to configure logging level
Add the ability to configure the logging level for a given invocation
of the Lambda. We use an environment variable to also highlight how a
Lambda might be configured to run with a specific configuration. This
environment variable will change the logging level for a given
invocation of the Lambda and then reset it to the previous level.