Restructure Lambda's Python code
Convert the Python code for the Lambda into a Python package. This is
done to enable native Python testing as well as to provide a
consolidated namespace for any of the Python code stored in the Lambda
project. This code should remain as minimal as possible, but this at
least provides better organization and testing.

Note that there is no `setup.py` or equivalent, and that any Lambda
dependecies are _still_ managed by the `pipenv` configurations stored
in the `runtimes/` directory. This change is being made solely to make
native Python testing possible and to provide a namespace for any
Lambda Python code.