Restructure Lambda to target a single runtime
Since Lambdas are essentially standalone serverless applications it
makes sense to distill their runtime support down to a single version.
There is no need to try and support multiple runtimes in any of the
Python code stored in the project. This simplifies builds and makes it
easier to integrate testing by only tracking a single set of pipenv
configuration files.