Target a single AWS Lambda Python runtime
Instead of building for multiple Lambda runtimes it makes sense to just
build targeting a single runtime. When deployed the infrastructure
will just need to use the same runtime as supported by the Lambda
configuration. Thus we drop support for all but the latest runtime that
is specified in the configuration.